seperate grid js from robot-simulator
