+++ index.html
@@ -7,6 +7,7 @@
     <title>Robot Test</title>
     <script type="module" src="robot-simulator.js"></script>
     <script type="module" src="robot-simulator.spec.js"></script>
+    <script type="module" src="grid.js"></script>
     <link rel="stylesheet" href="style.css" />
   </head>
   <body>
@@ -31,7 +32,6 @@
     </p>
     <p>The letter-string "RAALAL" means:</p>
     <p>Turn right Advance twice Turn left Advance once Turn left yet again</p>
-    <span id="standard"></span>
-    <div id="grid"></div>
+    <div id="grid"><span id="zero"></span></div>
   </body>
 </html>
