+++ style.css
@@ -1,29 +1,31 @@
 * {
   box-sizing: border-box;
+  margin: 0;
+  padding: 0;
 }
 html {
   font-size: 20px;
 }
 body {
   font-family: system-ui, sans-serif;
+  font-size: 1rem;
 }
-#standard {
+#zero {
   display: block;
   position: absolute;
   width: 1rem;
   height: 1rem;
-  bottom: 15.25rem;
-  left: 14.9rem;
+  left: 19.5rem;
+  top: 19.5rem;
   background-color: #294029;
-  border-top: 3px solid #15dae3;
-  /* border-right: 3px solid #15dae3;
+  /* border-top: 3px solid #15dae3;
+  border-right: 3px solid #15dae3;
   border-bottom: 3px solid #15dae3;
   border-left: 3px solid #15dae3; */
 }
 #grid {
   position: absolute;
-  width: 30rem;
-  height: 30rem;
+  width: 40rem;
+  height: 40rem;
   background: url(./grid.svg);
-  z-index: -1;
 }
